+++ include/seqan3/alignment/matrix/detail/affine_cell_proxy.hpp
@@ -17,7 +17,7 @@
 #include <type_traits>
 
 #include <seqan3/alignment/matrix/trace_directions.hpp>
-#include <seqan3/core/concept/core_language.hpp>
+#include <seqan3/utility/detail/exposition_only_concept.hpp>
 #include <seqan3/core/detail/empty_type.hpp>
 #include <seqan3/utility/simd/concept.hpp>
 #include <seqan3/utility/tuple_utility.hpp>
