+++ include/seqan3/alignment/matrix/detail/alignment_score_matrix_one_column_base.hpp
@@ -16,7 +16,7 @@
 #include <utility>
 #include <vector>
 
-#include <seqan3/core/concept/core_language.hpp>
+#include <seqan3/utility/detail/exposition_only_concept.hpp>
 #include <seqan3/range/container/aligned_allocator.hpp>
 #include <seqan3/utility/simd/concept.hpp>
