+++ include/seqan3/alignment/matrix/detail/combined_score_and_trace_matrix.hpp
@@ -16,7 +16,7 @@
 
 #include <seqan3/alignment/matrix/detail/affine_cell_proxy.hpp>
 #include <seqan3/alignment/matrix/detail/matrix_coordinate.hpp>
-#include <seqan3/core/concept/core_language.hpp>
+#include <seqan3/utility/detail/exposition_only_concept.hpp>
 #include <seqan3/core/range/type_traits.hpp>
 #include <seqan3/range/views/zip.hpp>
