+++ include/seqan3/argument_parser/validators.hpp
@@ -21,7 +21,7 @@
 #include <sstream>
 
 #include <seqan3/argument_parser/exceptions.hpp>
-#include <seqan3/core/concept/core_language.hpp>
+#include <seqan3/utility/detail/exposition_only_concept.hpp>
 #include <seqan3/core/debug_stream/detail/to_string.hpp>
 #include <seqan3/core/detail/debug_stream_range.hpp>
 #include <seqan3/io/detail/misc.hpp>
