+++ include/seqan3/io/detail/magic_header.hpp
@@ -18,8 +18,8 @@
 #include <seqan3/std/type_traits>
 #include <vector>
 
-#include <seqan3/core/concept/core_language.hpp>
 #include <seqan3/core/detail/template_inspection.hpp>
+#include <seqan3/utility/detail/exposition_only_concept.hpp>
 #include <seqan3/utility/detail/to_little_endian.hpp>
 #include <seqan3/utility/type_pack/traits.hpp>
