+++ include/seqan3/range/views/convert.hpp
@@ -12,7 +12,7 @@
 
 #pragma once
 
-#include <seqan3/core/concept/core_language.hpp>
+#include <seqan3/utility/detail/exposition_only_concept.hpp>
 #include <seqan3/std/concepts>
 #include <seqan3/std/ranges>
